+++ CHANGES.rst
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_,
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
+[0.0.28] - 2023-06-30
+=====================
+
+Fixed
+*****
+
+- A template variable was misnamed.
+
 [0.0.27] - 2023-06-29
 =====================